requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch
torchvision
timm
kaggle
kagglehub
scikit-learn
matplotlib
seaborn
pandas
pillow
numpy
